+++ src/Discovery.php
@@ -43,9 +43,7 @@ class Discovery
     {
         $this->ydb = $ydb;
 
-        $this->client = new ServiceClient($ydb->endpoint(), [
-            'credentials' => $ydb->iam()->getCredentials(),
-        ]);
+        $this->client = new ServiceClient($ydb->endpoint(), $ydb->grpcOpts());
 
         $this->meta = $ydb->meta();