+++ src/Operations.php
@@ -38,9 +38,7 @@ class Operations
     {
         $this->ydb = $ydb;
 
-        $this->client = new ServiceClient($ydb->endpoint(), [
-            'credentials' => $ydb->iam()->getCredentials(),
-        ]);
+        $this->client = new ServiceClient($ydb->endpoint(), $ydb->grpcOpts());
 
         $this->meta = $ydb->meta();